requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
MetaTrader5>=5.0.45
openai>=1.30.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.2.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
pydantic>=2.6.0
numpy>=1.26.0
